添加 ColorBuffer 和 DepthBuffer；gpu buffer 默认创建为 Common State ; 支持 Copy command list 初始化 dynamic 和 非 dynamic gpu buffer

+++ DashProject/DashCore/Src/Graphics/PipelineStateObject.cpp
@@ -126,7 +126,7 @@ namespace Dash
 		ASSERT(shader && shader->GetShaderStage() != EShaderStage::Compute);
 
 		const void* shaderData = shader->GetCompiledShader().Data;
-		uint64_t shaderSize = shader->GetCompiledShader().Size;
+		uint32_t shaderSize = shader->GetCompiledShader().Size;
 
 		switch (shader->GetShaderStage())
 		{
@@ -232,7 +232,7 @@ namespace Dash
 		ASSERT(shader && shader->GetShaderStage() == EShaderStage::Compute);
 
 		const void* shaderData = shader->GetCompiledShader().Data;
-		uint64_t shaderSize = shader->GetCompiledShader().Size;
+		uint32_t shaderSize = shader->GetCompiledShader().Size;
 
 		SetComputeShader(shaderData, shaderSize);
 	}
